tests: Make effective config check work for container runs
In https://github.com/linux-system-roles/tox-lsr/pull/191 we added
`ansible_become=false` to our container (podman/buildah) inventories.
This disables `become_user`, and thus the test was silently running as
root and thus failing. This isn't important for the stat tests (they are
fine to run as root), but `ssh` must actually run as the desired target
user.

This is just a test, so we can call `su` directly.