fix: Rewrite template for Ansible 2.19 compatibility
Cause: Ansible 2.19 changed behaviour: Any value which is "none" gets
rendered as literal "None" string instead of "not rendered"/"empty
string.

Consequence: The generated config file was completely broken with
Ansible 2.19.

Fix: Add explicit `is not none` guards. Also refactor the template with
an `ssh_options` list to be less repetitive.

Thanks to Claude-Sonnet 3.7 LLVM which heavily assisted with this
change.